fix: make sure common search functions are exported from search.js itself
