+++ server/lib/search.js
@@ -11,7 +11,6 @@ import models, { Op, sequelize } from '../models';
 
 import { floatAmountToCents } from './math';
 import RateLimit, { ONE_HOUR_IN_SECONDS } from './rate-limit';
-import { sanitizeSearchTermForILike, searchTermToTsVector, trimSearchTerm } from './utils';
 
 // Returned when there's no result for a search
 const EMPTY_SEARCH_RESULT = [[], 0];
@@ -58,6 +57,29 @@ export const searchCollectivesByEmail = async (email, user, offset = 0, limit = 
   );
 
   return [collectives, get(collectives[0], 'dataValues.__total__', 0)];
+};
+
+/**
+ * Turn a search string into a TS vector using 'OR' operator.
+ *
+ * Ex: "open potatoes" => "open|potatoes"
+ */
+export const searchTermToTsVector = term => {
+  return term.replace(/\s+/g, '|');
+};
+
+/**
+ * Trim leading/trailing spaces and remove multiple spaces from the string
+ */
+export const trimSearchTerm = term => {
+  return term?.trim().replace(/\s+/g, ' ');
+};
+
+/**
+ * Removes special ILIKE characters like `%
+ */
+export const sanitizeSearchTermForILike = term => {
+  return term.replace(/(_|%|\\)/g, '\\$1');
 };
 
 /**
